@@ -29,12 +29,15 @@
     "node-sass": "^4.7.2",
     "popper.js": "^1.12.5",
     "postcss-loader": "^2.1.1",
+    "pug": "^2.0.3",
+    "pug-html-loader": "^1.1.5",
+    "pug-loader": "^2.4.0",
     "resolve-url-loader": "^2.3.0",
     "sass-loader": "^6.0.7",
     "style-loader": "^0.20.3",
     "twig-loader": "^0.3.1",
     "url-loader": "^1.0.1",
-    "webpack": "^2.3.2",
+    "webpack": "^2.6.0",
     "webpack-dev-server": "^2.4.2"
   },
   "dependencies": {
@@ -43,6 +46,8 @@
     "@fortawesome/fontawesome-free-solid": "^5.0.8",
     "animate.css": "^3.6.1",
     "bootstrap-loader": "^2.2.0",
-    "jquery": "^3.3.1"
+    "jquery": "^3.3.1",
+    "lightbox2": "^2.10.0",
+    "vue": "^2.5.16"
   }
 }
