@@ -1,5 +1,5 @@
 import '../style/main.scss';
-
+const newsSet = require('./news-set');
 
 $(document).ready(function() {
   $("body").css("display", "block");
@@ -7,4 +7,17 @@ $(document).ready(function() {
   $("#patreons").on("click", function(){
     window.location = "patreon.html"
   });
+
+  $.ajax({
+    type: "get",
+    url: "http://upnp.ga/api/news",
+    success: function (response) {
+      var newsList = newsSet.getNews(response);
+      $("#news-block").append(newsList);
+    },
+    contentType: false,
+    cache: false,
+    processData: false,
+  });
+
 })
