@@ -3,7 +3,9 @@ import "../style/main_sn.scss";
 $(document).ready(function(){
   $("body").css("display", "block");
   $("#logo").on('click', function(){
-    window.location = '/';
+    var url = $(this).attr("data-url");
+    // console.log(url)
+    window.location = url;
   })
   
 });
