@@ -4,9 +4,11 @@ import "./additional-methods.min.js";
 
 $(document).ready(function () {
   $("body").css("display", "block");
-  $("#logo").on('click', function () {
-    window.location = '/';
-  });
+  $("#logo").on('click', function(){
+    var url = $(this).attr("data-url");
+    // console.log(url)
+    window.location = url;
+  })
 
   $("#volontieer-form").submit(function (e) {
     e.preventDefault();
@@ -16,7 +18,9 @@ $(document).ready(function () {
       form[el.name] = el.value;
       return form;
     }, {});
-    console.log(data);
+    
+    var nedeljni_sati =$('input.nedeljni_sati:checked');
+    console.log(nedeljni_sati);
     $.ajax({
       type: "POST",
       url: "http://upnp.ga/volountieer/create",
@@ -32,7 +36,7 @@ $(document).ready(function () {
   })
   $("#volontieer-form").validate({
     rules: {
-      name: {
+      ime_prezime: {
         required: true
       },
       datum: {
